remove ArticleBreadcrumbGenerator extension from project-base
- the parent class now implements DomainBreadcrumbGeneratorInterface
