--- project-base/app/src/Model/Article/ArticleBreadcrumbGenerator.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Model\Article;
-
-use Shopsys\FrameworkBundle\Component\Breadcrumb\DomainBreadcrumbGeneratorInterface;
-use Shopsys\FrameworkBundle\Model\Article\ArticleBreadcrumbGenerator as BaseArticleBreadcrumbGenerator;
-
-class ArticleBreadcrumbGenerator extends BaseArticleBreadcrumbGenerator implements DomainBreadcrumbGeneratorInterface
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function getBreadcrumbItemsOnDomain(
-        int $domainId,
-        string $routeName,
-        array $routeParameters = [],
-        ?string $locale = null,
-    ): array {
-        return $this->getBreadcrumbItems($routeName, $routeParameters);
-    }
-}
